Dev/config spring security

+++ refoods/src/main/java/com/projeto/ReFood/repository/ContactRepository.java
@@ -3,4 +3,4 @@ package com.projeto.ReFood.repository;
 import com.projeto.ReFood.model.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactRepository extends JpaRepository<Contact, Integer> {}
+public interface ContactRepository extends JpaRepository<Contact, Long> {}
