+++ refoods/src/main/java/com/projeto/ReFood/repository/NotificationRepository.java
@@ -4,4 +4,6 @@ import com.projeto.ReFood.model.Notification;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {}
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
+
+}
