+++ refoods/src/main/java/com/projeto/ReFood/repository/ReviewRepository.java
@@ -4,4 +4,4 @@ import com.projeto.ReFood.model.Review;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {}
+public interface ReviewRepository extends JpaRepository<Review, Long> {}
