+++ refoods/src/main/java/com/projeto/ReFood/repository/TransactionRepository.java
@@ -4,4 +4,4 @@ import com.projeto.ReFood.model.Transaction;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {}
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {}
